bypassing checks when running make distcheck

+++ test/gils/stop02.sh
@@ -1,21 +1,27 @@
 #!/bin/sh
-# $Id: stop02.sh,v 1.11 2005-01-03 09:19:26 adam Exp $
+# $Id: stop02.sh,v 1.12 2006-04-27 10:52:26 marc Exp $
 # test start and stop of the server with -S
 
-pp=${srcdir:-"."}
+srcdir=${srcdir:-"."}
+
+if [ "$srcdir" != "." ]
+    then
+    echo "Jumping over test"
+    exit 0
+fi
 
 LOG=stop02.log
 
 rm -f $LOG
 echo "initializing" >>$LOG
 test -d reg || mkdir reg
-../../index/zebraidx -l $LOG -c $pp/zebra1.cfg init || exit 1
+../../index/zebraidx -l $LOG -c $srcdir/zebra1.cfg init || exit 1
 
 #create a base to test on
-../../index/zebraidx -l $LOG -c $pp/zebra1.cfg update records  || exit 1
+../../index/zebraidx -l $LOG -c $srcdir/zebra1.cfg update records  || exit 1
 
 echo "Starting server with -S (static)..." >>$LOG
-../../index/zebrasrv -D -p z.pid -S -c $pp/zebra1.cfg -l $LOG unix:socket
+../../index/zebrasrv -D -p z.pid -S -c $srcdir/zebra1.cfg -l $LOG unix:socket
 echo "  checking that it runs... " >>$LOG
 test -f z.pid || exit 1
 PID=`cat z.pid`
