+++ test/gils/stop03.sh
@@ -1,21 +1,27 @@
 #!/bin/sh
-# $Id: stop03.sh,v 1.11 2005-01-03 12:08:04 adam Exp $
+# $Id: stop03.sh,v 1.12 2006-04-27 10:52:26 marc Exp $
 # test start and stop of the threaded server (-T)
 
-pp=${srcdir:-"."}
+srcdir=${srcdir:-"."}
+
+if [ "$srcdir" != "." ]
+    then
+    echo "Jumping over test"
+    exit 0
+fi
 
 LOG=stop03.log
 
 rm -f $LOG
 echo "initializing" >>$LOG
 mkdir -p reg
-../../index/zebraidx -l $LOG -c $pp/zebra1.cfg init || exit 1
+../../index/zebraidx -l $LOG -c $srcdir/zebra1.cfg init || exit 1
 
 #create a base to test on
-../../index/zebraidx -l $LOG -c $pp/zebra1.cfg update records  || exit 1
+../../index/zebraidx -l $LOG -c $srcdir/zebra1.cfg update records  || exit 1
 
 echo "Starting server with -T (threaded)..." >>$LOG
-../../index/zebrasrv -D -p z.pid -T -c $pp/zebra1.cfg -l $LOG unix:socket 2>out
+../../index/zebrasrv -D -p z.pid -T -c $srcdir/zebra1.cfg -l $LOG unix:socket 2>out
 
 if grep 'not available' out >/dev/null; then
     test -f z.pid && rm -f z.pid
