+++ test/gils/test1.sh
@@ -1,21 +1,39 @@
 #!/bin/sh
-# $Id: test1.sh,v 1.13 2005-01-03 12:08:04 adam Exp $
+# $Id: test1.sh,v 1.14 2006-04-27 10:52:26 marc Exp $
 
-pp=${srcdir:-"."}
+srcdir=${srcdir:-"."}
+
+if [ "$srcdir" != "." ]
+    then
+    echo "Jumping over test"
+    exit 0
+fi
+
 
 LOG=test1.log
 
 rm -f $LOG
 echo  "initializing..." >>$LOG
 test -d reg || mkdir reg
-rm -f $pp/records/esdd000[12].grs # these should not be here, will be created later
-../../index/zebraidx -l $LOG -c $pp/zebra1.cfg init || exit 1
+
+# these should not be here, will be created later
+$srcdir/cleanrecords.sh
+#if [ -f $srcdir/records/esdd0001.grs ] 
+#    then
+#    rm -f $srcdir/records/esdd0001.grs
+#fi
+#if [ -f $srcdir/records/esdd0002.grs ] 
+#    then
+#    rm -f $srcdir/records/esdd0002.grs
+#fi
+
+../../index/zebraidx -l $LOG -c $srcdir/zebra1.cfg init || exit 1
 
 echo "updating..." >>$LOG
-../../index/zebraidx -l $LOG -c $pp/zebra1.cfg update $pp/records  || exit 1
+../../index/zebraidx -l $LOG -c $srcdir/zebra1.cfg update $srcdir/records  || exit 1
 
 echo "starting server..." >>$LOG
-../../index/zebrasrv -D -p z.pid -S -c $pp/zebra1.cfg -l $LOG unix:socket
+../../index/zebrasrv -D -p z.pid -S -c $srcdir/zebra1.cfg -l $LOG unix:socket
 
 echo "checking it runs..." >>$LOG
 test -f z.pid || exit 1
@@ -37,7 +55,7 @@ echo "search 4..." >>$LOG
 grep "^Result count: 9$" log >/dev/null || exit 1
 
 echo "reindexing..." >>$LOG
-../../index/zebraidx -l $LOG -c $pp/zebra1.cfg update $pp/records || exit 1
+../../index/zebraidx -l $LOG -c $srcdir/zebra1.cfg update $srcdir/records || exit 1
 
 echo "search 5..." >>$LOG
 ../api/testclient unix:socket "@attr 1=4 utah" > log || exit 1
